+++ Sketch/MiniCNC/MiniCNC/MyControl.cpp
@@ -86,8 +86,8 @@ void CMyControl::Init()
 {
 	CSingleton<CConfigEeprom>::GetInstance()->Init(sizeof(CConfigEeprom::SCNCEeprom), &_eepromFlash, EPROM_SIGNATURE);
 
-#ifdef DISABLELEDBLINK
-	DisableBlinkLed();
+#ifdef BLINK_LED
+	CHAL::pinModeOutput(BLINK_LED);
 #endif
 
 	super::Init();
@@ -207,6 +207,15 @@ void CMyControl::Poll()
 		Lcd.Diagnostic(F("LCD Hold"));
 #endif
 	}
+#ifdef BLINK_LED
+
+	uint32_t time = millis();
+	if (_timeBlink < time)
+	{
+		HALFastdigitalWrite(BLINK_LED, CHAL::digitalRead(BLINK_LED) == HIGH ? LOW : HIGH);
+		_timeBlink = time + BLINK_TIMEOUT;
+	}
+#endif
 }
 
 ////////////////////////////////////////////////////////////
