+++ Sketch/MiniCNC/MiniCNC/MyControl.h
@@ -67,6 +67,10 @@ protected:
 	
 private:
 
+#ifdef BLINK_LED
+	uint32_t _timeBlink=0;
+#endif
+
 	static const CConfigEeprom::SCNCEeprom _eepromFlash;
 
 	ControlData _data;
