+++ Sketch/libraries/StepperLib/src/Steppers/StepperCNCShield_pins.h
@@ -96,7 +96,7 @@
 
 #define CNCSHIELD_SPINDLE_DIGITAL_ON		LOW
 #define CNCSHIELD_SPINDLE_DIGITAL_OFF		HIGH
-#define CNCSHIELD_spindle_ANALOG_MAX		255
+#define CNCSHIELD_SPINDLE_ANALOG_MAX		255
 
 #define CNCSHIELD_SPINDLE_DIR_CLW	LOW
 #define CNCSHIELD_SPINDLE_DIR_CCLW	HIGH
