+++ VS/Arduino.VC/MiniCNC/MiniCNC.vcxproj
@@ -86,18 +86,18 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Configuration.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_CNC3325.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_CNCShield3x.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_Custom.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_Default.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_Esp32C6.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_L298N.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_SMC800.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationMachine_TB6560.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationStepper_CNCShield3x.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationStepper_L298N.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationStepper_SMC800.h" />
-    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\ConfigurationStepper_TB6560.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_CNC3325.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_CNCShield3x.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_Custom.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_Default.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_Esp32C6.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_L298N.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_SMC800.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Machine_TB6560.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Stepper_CNCShield3x.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Stepper_L298N.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Stepper_SMC800.h" />
+    <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\Machines\Stepper_TB6560.h" />
     <ClInclude Include="..\..\..\Sketch\MiniCNC\MiniCNC\MyControl.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
